refactor(be): change authuser interface to class
