--- backend/src/auth/interface/authenticated-user.interface.ts
@@ -1,7 +1,0 @@
-import { Role } from '@prisma/client'
-
-export interface AuthenticatedUser {
-  id: number
-  username: string
-  role?: Role
-}
